+++ src/cli/server/mod.rs
@@ -1,10 +1,9 @@
 pub mod http_1;
 pub mod http_2;
-#[allow(clippy::module_inception)]
-pub mod server;
+pub mod http_server;
 pub mod server_config;
 
-pub use server::Server;
+pub use http_server::Server;
 
 use self::server_config::ServerConfig;
